requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
et-xmlfile==1.1.0
idna==3.10
openpyxl==3.1.5
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
soupsieve==2.6
urllib3==2.2.3
